test: isolate AE2 configurable fixtures
