+++ projects/typescript-tests/src/ae2_formation_plane.ts
@@ -1,3 +1,47 @@
-import { testConfigurableObject } from "./ae2_configurable_object";
+import type { AE2CableAPI, AE2FormationPlaneObject } from "@siredvin/typed-peripheral-unlimitedperipheralworks/integrations/ae2Objects";
+import type { InventoryViewAPI } from "@siredvin/typed-peripheral-api/inventory_view";
 
-testConfigurableObject("formation_plane");
+/** @noSelf **/
+interface TestApi { ok(marker?: string): void; }
+declare const test: TestApi;
+
+const check = (value: unknown, message: string): void => { if (!value) throw message; };
+const fails = (action: (this: void) => void, message: string): void => {
+    const [ok] = pcall(action);
+    check(!ok, message);
+};
+const item = (name: string) => ({ type: "item" as const, name });
+const fluid = (name: string) => ({ type: "fluid" as const, name });
+const sides = ["top", "bottom", "left", "right", "front", "back"];
+let inventory: InventoryViewAPI | undefined;
+let inventoryName = "";
+let object: any;
+for (let attempt = 0; attempt < 100; attempt++) {
+    for (const side of sides) {
+        const wrapped = peripheral.wrap(side) as any;
+        if (!wrapped) continue;
+        if (wrapped.getSide) {
+            const [device] = (wrapped as AE2CableAPI).getSide("south");
+            if (device) object = device;
+        } else if (wrapped.listStock || wrapped.listPatterns) {
+            object = wrapped;
+        } else if (wrapped.list) {
+            inventory = wrapped as InventoryViewAPI;
+            inventoryName = side;
+        }
+    }
+    if (inventory && object) break;
+    sleep(0.05);
+}
+check(inventory, "Fixture inventory did not become available");
+check(object, "Fixture configurable object did not become available");
+const itemInventory = inventory as InventoryViewAPI;
+
+check(object.getDeviceType() === "formation_plane", `Expected formation_plane, got ${object.getDeviceType()}`);
+    const target = object as AE2FormationPlaneObject;
+    target.setFilter(1, item("minecraft:stone"));
+    target.setPriority(4);
+    target.setPlaceBlocks(false);
+    check(target.getFilter(1)?.name === "minecraft:stone" && target.getPriority() === 4 && !target.shouldPlaceBlocks(), "Formation Plane settings did not apply");
+test.ok();
+
