+++ projects/typescript-tests/src/ae2_pattern_provider.ts
@@ -1,3 +1,53 @@
-import { testConfigurableObject } from "./ae2_configurable_object";
+import type { AE2CableAPI, AE2PatternProviderPeripheral } from "@siredvin/typed-peripheral-unlimitedperipheralworks/integrations/ae2Objects";
+import type { InventoryViewAPI } from "@siredvin/typed-peripheral-api/inventory_view";
 
-testConfigurableObject("pattern_provider");
+/** @noSelf **/
+interface TestApi { ok(marker?: string): void; }
+declare const test: TestApi;
+
+const check = (value: unknown, message: string): void => { if (!value) throw message; };
+const fails = (action: (this: void) => void, message: string): void => {
+    const [ok] = pcall(action);
+    check(!ok, message);
+};
+const item = (name: string) => ({ type: "item" as const, name });
+const fluid = (name: string) => ({ type: "fluid" as const, name });
+const sides = ["top", "bottom", "left", "right", "front", "back"];
+let inventory: InventoryViewAPI | undefined;
+let inventoryName = "";
+let object: any;
+for (let attempt = 0; attempt < 100; attempt++) {
+    for (const side of sides) {
+        const wrapped = peripheral.wrap(side) as any;
+        if (!wrapped) continue;
+        if (wrapped.getSide) {
+            const [device] = (wrapped as AE2CableAPI).getSide("south");
+            if (device) object = device;
+        } else if (wrapped.listStock || wrapped.listPatterns) {
+            object = wrapped;
+        } else if (wrapped.list) {
+            inventory = wrapped as InventoryViewAPI;
+            inventoryName = side;
+        }
+    }
+    if (inventory && object) break;
+    sleep(0.05);
+}
+check(inventory, "Fixture inventory did not become available");
+check(object, "Fixture configurable object did not become available");
+const itemInventory = inventory as InventoryViewAPI;
+
+check(object.getDeviceType() === "pattern_provider", `Expected pattern_provider, got ${object.getDeviceType()}`);
+    const target = object as AE2PatternProviderPeripheral;
+    target.setPriority(9);
+    target.setBlocking(true);
+    target.setVisibleInPatternAccessTerminal(false);
+    target.setPatternLockMode("lock_while_high");
+    target.setPushDirection("east");
+    check(target.getPriority() === 9 && target.isBlocking() && !target.isVisibleInPatternAccessTerminal() &&
+        target.getPatternLockMode() === "lock_while_high" && target.getPushDirection() === "east", "pattern provider settings did not apply");
+    check(target.pullPattern(inventoryName, 6, 1, 1) === 1 && target.getPattern(1) !== null, "encoded pattern did not transfer into the provider");
+    check(target.pullPattern(inventoryName, 5, 1, 2) === 0 && itemInventory.list()[5]?.name === "minecraft:stone", "non-pattern transfer mutated its source");
+    check(target.pushPattern(inventoryName, 1, 1, 12) === 1, "encoded pattern did not transfer out of the provider");
+test.ok();
+
